issue #1116: move physicsSprite into the singal dir
